(BKR-358) split up monolithic test file
